refactor: 검색 기능 성능 개선을 위한 MDX 로직 분리 #130
